requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
slowapi
pydantic[email]
cryptography
httpx
